Support larger ranges for algorithms (#9503)
* Specify IndexType to default rangepolicies internally used in algorithms
* Replace with Iterator::difference_type
* Added few missing headers
* Apply suggetsions from review and added few large_mem tests
* Fix in SortImpl min_max_functor's operator

---------

+++ algorithms/src/sorting/Kokkos_BinOpsPublicAPI.hpp
@@ -5,6 +5,7 @@
 #define KOKKOS_BIN_OPS_PUBLIC_API_HPP_
 
 #include <Kokkos_Macros.hpp>
+#include <cstdint>
 #include <type_traits>
 
 namespace Kokkos {
@@ -37,7 +38,7 @@ struct BinOp1D {
 
   // Determine bin index from key value
   template <class ViewType>
-  KOKKOS_INLINE_FUNCTION int bin(ViewType& keys, const int& i) const {
+  KOKKOS_INLINE_FUNCTION int bin(ViewType& keys, const int64_t& i) const {
     return static_cast<int>(mul_ * (static_cast<double>(keys(i)) - min_));
   }
 
@@ -78,7 +79,7 @@ struct BinOp3D {
   }
 
   template <class ViewType>
-  KOKKOS_INLINE_FUNCTION int bin(ViewType& keys, const int& i) const {
+  KOKKOS_INLINE_FUNCTION int bin(ViewType& keys, const int64_t& i) const {
     return int((((int(mul_[0] * (keys(i, 0) - min_[0])) * max_bins_[1]) +
                  int(mul_[1] * (keys(i, 1) - min_[1]))) *
                 max_bins_[2]) +
